+++ .ci/build.sh
@@ -7,7 +7,6 @@ echo "-- ${0} start..."
 echo "-- Building docker image for production"
 
 buildDocker \
-	--progress=plain \
 	--file .ci/Dockerfile \
 	--tag "${IMAGE_PATH}:${VERSION}" \
 	.