+++ .ci/run.sh
@@ -10,9 +10,6 @@ docker run \
   -t \
 	--rm \
 	--name=${APP_NAME} \
-	-e TURBO_TEAM="${TURBO_TEAM}" \
-	-e TURBO_TOKEN="${TURBO_TOKEN}" \
-	-e TURBO_REMOTE_ONLY=true \
 	${IMAGE_PATH}:${VERSION} \
 	${@}