+++ packages/eslint-config-react/.eslintrc.js
@@ -1,3 +1,0 @@
-module.exports = {
-    extends: ['@wesp-up/eslint-config'],
-};