+++ packages/eslint-config-react/CHANGELOG.md
@@ -1,11 +1,26 @@
-# @wesp-up/eslint-config
+# @wesp-up/eslint-config-react
+
+## 2.0.0 (2024-07-27)
+
+### Major Changes
+
+Update ESLint to v9
+
+- This required a lot of changes and some configs/plugins still haven't updated yet, so will have to wait until they do (e.g. eslint-plugin-jsx-a11y).
+- Removed airbnb configs in favor of using default recommended rules to greatly simplify. Recommended rule sets have also gotten really good over time.
+- Removed prettier plugin and config in favor of using prettier directly, which is their recommendation.
+
+### Patch Changes
+
+- Updated dependencies
+  - @wesp-up/eslint-config@2.0.0
 
 ## 1.0.1 (2024-02-14)
 
 ### Patch Changes
 
 - Updated dependencies
-    - @wesp-up/eslint-config@1.0.1
+  - @wesp-up/eslint-config@1.0.1
 
 ## 1.0.0 (2023-11-25)