+++ packages/tsconfig/CHANGELOG.md
@@ -1,8 +1,14 @@
 # @wesp-up/tsconfig
 
+## 2.0.0 (2024-07-27)
+
+### Major Changes
+
+- - Removed vitest types from the base config since the types are not always present.
+  - Changed the `node.json` config from `"module": "CommonJS"` to `"module": "ESNext"` due to the major shift to ES Modules.
+
 ## 1.0.0 (2023-11-25)
 
 ### Major Changes
 
 - Initial release
-